+++ backend/backend/settings/api_events/common.py
@@ -7,8 +7,8 @@ INSTALLED_APPS = [
     "django.contrib.contenttypes",
     "django_celery_results",
     "rest_framework",
-    "localrep",
-    "localsync",
+    "api",
+    "api_events",
     "substrapp",
     "users",
     "organization",
