+++ backend/substrapp/models/__init__.py
@@ -5,7 +5,7 @@ from .datamanager import DataManager
 from .datasample import DataSample
 from .image_entrypoint import ImageEntrypoint
 from .model import Model
-from .worker_events import WorkerLastEvent
+from .worker_last_event import WorkerLastEvent
 
 __all__ = [
     "DataSample",
