+++ backend/worker_events/handler_compute_engine.py
