+++ charts/substra-backend/templates/rbac.yaml
@@ -52,7 +52,7 @@ roleRef:
   apiGroup: rbac.authorization.k8s.io
 {{- end }}
 
-{{- if .Values.events.serviceAccount.create }}
+{{- if .Values.worker_events.serviceAccount.create }}
 ---
 apiVersion: v1
 kind: ServiceAccount
@@ -62,12 +62,12 @@ metadata:
     {{ include "substra.labels" . | nindent 4 }}
     app.kubernetes.io/name: {{ template "substra.name" . }}
 {{- end -}}
-{{- if .Values.events.rbac.create }}
+{{- if .Values.worker_events.rbac.create }}
 ---
 kind: Role
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
-  name: {{ template "substra.fullname" . }}-event
+  name: {{ template "substra.fullname" . }}-worker-event
   labels:
     {{ include "substra.labels" . | nindent 4 }}
     app.kubernetes.io/name: {{ template "substra.name" . }}
@@ -79,15 +79,54 @@ rules:
 kind: RoleBinding
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
-  name: {{ template "substra.fullname" . }}-event
+  name: {{ template "substra.fullname" . }}worker-event
   labels:
     {{ include "substra.labels" . | nindent 4 }}
     app.kubernetes.io/name: {{ template "substra.name" . }}
 subjects:
   - kind: ServiceAccount
-    name: {{ include "substra.events.serviceAccountName" . }}
+    name: {{ include "substra.worker_events.serviceAccountName" . }}
 roleRef:
   kind: Role
-  name: {{ template "substra.fullname" . }}-event
+  name: {{ template "substra.fullname" . }}worker-event
+  apiGroup: rbac.authorization.k8s.io
+{{- end }}
+{{- if .Values.api_events.serviceAccount.create }}
+---
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: {{ template "substra.fullname" . }}api-event
+  labels:
+    {{ include "substra.labels" . | nindent 4 }}
+    app.kubernetes.io/name: {{ template "substra.name" . }}
+{{- end -}}
+{{- if .Values.api_events.rbac.create }}
+---
+kind: Role
+apiVersion: rbac.authorization.k8s.io/v1
+metadata:
+  name: {{ template "substra.fullname" . }}api-event
+  labels:
+    {{ include "substra.labels" . | nindent 4 }}
+    app.kubernetes.io/name: {{ template "substra.name" . }}
+rules:
+  - apiGroups: ["apps"]
+    resources: ["statefulsets/scale"]
+    verbs: ["get"]
+---
+kind: RoleBinding
+apiVersion: rbac.authorization.k8s.io/v1
+metadata:
+  name: {{ template "substra.fullname" . }}api-event
+  labels:
+    {{ include "substra.labels" . | nindent 4 }}
+    app.kubernetes.io/name: {{ template "substra.name" . }}
+subjects:
+  - kind: ServiceAccount
+    name: {{ include "substra.api_events.serviceAccountName" . }}
+roleRef:
+  kind: Role
+  name: {{ template "substra.fullname" . }}api-event
   apiGroup: rbac.authorization.k8s.io
 {{- end }}
