+++ charts/substra-backend/values.yaml
@@ -449,17 +449,17 @@ containerRegistry:
   ##
   prepopulate: []
 
-## @section Event app settings
+## @section Worker event app settings
 ##
-events:
-  ## @param events.enabled Enable event service
+worker_events:
+  ## @param worker_events.enabled Enable event service
   ##
   enabled: true
-  ## @param events.image.registry Substra event app image registry
-  ## @param events.image.repository Substra event app image repository
-  ## @param events.image.tag Substra event app image tag
-  ## @param events.image.pullPolicy Substra event app image pull policy
-  ## @param events.image.pullSecrets Specify image pull secrets
+  ## @param worker_events.image.registry Substra event app image registry
+  ## @param worker_events.image.repository Substra event app image repository
+  ## @param worker_events.image.tag Substra event app image tag
+  ## @param worker_events.image.pullPolicy Substra event app image pull policy
+  ## @param worker_events.image.pullSecrets Specify image pull secrets
   ##
   image:
     registry: ghcr.io
@@ -467,49 +467,49 @@ events:
     tag: 0.29.0
     pullPolicy: IfNotPresent
     pullSecrets: []
-  ## @param events.podSecurityContext.enabled Enable security context
-  ## @param events.podSecurityContext.runAsUser User ID for the pod
-  ## @param events.podSecurityContext.runAsGroup Group ID for the pod
-  ## @param events.podSecurityContext.fsGroup FileSystem group ID for the pod
+  ## @param worker_events.podSecurityContext.enabled Enable security context
+  ## @param worker_events.podSecurityContext.runAsUser User ID for the pod
+  ## @param worker_events.podSecurityContext.runAsGroup Group ID for the pod
+  ## @param worker_events.podSecurityContext.fsGroup FileSystem group ID for the pod
   ##
   podSecurityContext:
     enabled: true
     runAsUser: 1001
     runAsGroup: 1001
     fsGroup: 1001
-  ## @param events.nodeSelector Node labels for pod assignment
+  ## @param worker_events.nodeSelector Node labels for pod assignment
   ##
   nodeSelector: {}
-  ## @param events.tolerations Toleration labels for pod assignment
+  ## @param worker_events.tolerations Toleration labels for pod assignment
   ##
   tolerations: []
-  ## @param events.affinity Affinity settings for pod assignment
+  ## @param worker_events.affinity Affinity settings for pod assignment
   ##
   affinity: {}
-  ## @param events.rbac.create Create a role and service account for the event app
+  ## @param worker_events.rbac.create Create a role and service account for the event app
   ##
   rbac:
     create: true
   serviceAccount:
-    ## @param events.serviceAccount.create Create a service account for the event app
+    ## @param worker_events.serviceAccount.create Create a service account for the event app
     ##
     create: true
-    ## @param events.serviceAccount.name The name of the ServiceAccount to use
+    ## @param worker_events.serviceAccount.name The name of the ServiceAccount to use
     ## If not set and create is true, a name is generated using the substra.fullname template
     ##
     name: ""
 
-## @section Localsync app settings
+## @section Api event app settings
 ##
-localsync:
-  ## @param localsync.enabled Enable event service
+api_events:
+  ## @param api_events.enabled Enable event service
   ##
   enabled: true
-  ## @param localsync.image.registry Substra event app image registry
-  ## @param localsync.image.repository Substra event app image repository
-  ## @param localsync.image.tag Substra event app image tag
-  ## @param localsync.image.pullPolicy Substra event app image pull policy
-  ## @param localsync.image.pullSecrets Specify image pull secrets
+  ## @param api_events.image.registry Substra event app image registry
+  ## @param api_events.image.repository Substra event app image repository
+  ## @param api_events.image.tag Substra event app image tag
+  ## @param api_events.image.pullPolicy Substra event app image pull policy
+  ## @param api_events.image.pullSecrets Specify image pull secrets
   ##
   image:
     registry: gcr.io
@@ -517,34 +517,34 @@ localsync:
     tag: 0.28.0
     pullPolicy: IfNotPresent
     pullSecrets: []
-  ## @param localsync.podSecurityContext.enabled Enable security context
-  ## @param localsync.podSecurityContext.runAsUser User ID for the pod
-  ## @param localsync.podSecurityContext.runAsGroup Group ID for the pod
-  ## @param localsync.podSecurityContext.fsGroup FileSystem group ID for the pod
+  ## @param api_events.podSecurityContext.enabled Enable security context
+  ## @param api_events.podSecurityContext.runAsUser User ID for the pod
+  ## @param api_events.podSecurityContext.runAsGroup Group ID for the pod
+  ## @param api_events.podSecurityContext.fsGroup FileSystem group ID for the pod
   ##
   podSecurityContext:
     enabled: true
     runAsUser: 1001
     runAsGroup: 1001
     fsGroup: 1001
-  ## @param localsync.nodeSelector Node labels for pod assignment
+  ## @param api_events.nodeSelector Node labels for pod assignment
   ##
   nodeSelector: {}
-  ## @param localsync.tolerations Toleration labels for pod assignment
+  ## @param api_events.tolerations Toleration labels for pod assignment
   ##
   tolerations: []
-  ## @param localsync.affinity Affinity settings for pod assignment
+  ## @param api_events.affinity Affinity settings for pod assignment
   ##
   affinity: {}
-  ## @param localsync.rbac.create Create a role and service account for the event app
+  ## @param api_events.rbac.create Create a role and service account for the event app
   ##
   rbac:
     create: true
   serviceAccount:
-    ## @param localsync.serviceAccount.create Create a service account for the event app
+    ## @param api_events.serviceAccount.create Create a service account for the event app
     ##
     create: true
-    ## @param localsync.serviceAccount.name The name of the ServiceAccount to use
+    ## @param api_events.serviceAccount.name The name of the ServiceAccount to use
     ## If not set and create is true, a name is generated using the substra.fullname template
     ##
     name: ""
